+++ apps/router-e2e/__e2e__/server-middleware-async/app/index.tsx
@@ -1,0 +1,15 @@
+import { Text, View } from "react-native";
+
+export default function Index() {
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <Text testID="title">Index</Text>
+    </View>
+  );
+}